Resolve the merge.
In Eclipse, right-click on that conflicted Java file and select “Team” →
“merge tool”. You should see both versions side-by-side, with the two
distinct comment lines flagged as a problem.

In the left pane, choose one comment or remove it entirely. Save the
file.

In VSCode, view the list of conflicts in the git/source control view.
Click on the conflicted file name to open the difference editor.

Scroll down to where the conflict is displayed and select the version
you would like. Save the file.

Then add the changed file to the git stage. Because this newly changed
version of the file was saved after both the the conflicting versions,
git will accept this as a resolution of the conflict.

Commit and push your changes. This will conclude a successful merge of
the two branches.

Again, return to GitHub and examine the Branches. The change in the
Insights,Network page should be particularly obvious.